Signals (slots really) are now comperable.  Building two slots for the same function or same object and method will compare as expected.
Also added a -= function to Bu::List which works just like erase.

+++ src/stable/archive.h
@@ -129,7 +129,7 @@ namespace Bu
     private:
         Stream &rStream;
         uint32_t nNextID;
-        Hash<uint32_t,uint32_t> hPtrID;
+        Hash<uint32_t,ptrdiff_t> hPtrID;
         Hash<uint32_t,List<void **> > hPtrDest;
         Hash<Bu::String, Variant> hProps;
     };
